% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cdr_lifecycle_metrics.R
\name{calculate_cdr_lifecycle_metrics}
\alias{calculate_cdr_lifecycle_metrics}
\title{Function to Calculate and Summarize CDR Lifecycle Metrics}
\usage{
calculate_cdr_lifecycle_metrics(
  db_path,
  db_name,
  dat_file,
  scenario_list,
  region_list = NULL,
  output_path,
  output_type = c("csv", "list"),
  create_plots = TRUE,
  metric = "NCRR",
  ncol = 2,
  nrow = NULL,
  selected_years = NULL
)
}
\arguments{
\item{db_path}{Path to the GCAM database.}

\item{db_name}{Name of the GCAM database.}

\item{dat_file}{Name of the .dat file to load.}

\item{scenario_list}{List of scenarios to include.}

\item{region_list}{List of regions to filter, default is all regions.}

\item{output_path}{Path to save output files.}

\item{output_type}{Output format, either "csv" or "list". Default is "list".}

\item{create_plots}{Logical, if TRUE, generates plots for the results. Default is TRUE.}

\item{metric}{The metric to visualize in plots. One of "Total_Lifecycle_Emissions",
"Net_Removal_Efficiency", "NCRR". Default is "NCRR".}

\item{ncol}{Number of columns for the facets in the plots. Default is 2.}

\item{nrow}{Number of rows for the facets in the plots. Default is NULL.}

\item{selected_years}{For time-based visualizations, a vector of years to include, or NULL for all years.}
}
\value{
A list containing the calculated lifecycle metrics.
}
\description{
This function estimates lifecycle emissions metrics for Carbon Dioxide Removal (CDR) technologies
based on a supplied database and includes options for filtering, visualization, and saving results.
}
