% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCR.R
\name{calculate_ncrr}
\alias{calculate_ncrr}
\title{Calculate Net Carbon Removal Ratio (NCRR)}
\usage{
calculate_ncrr(data)
}
\arguments{
\item{data}{A data frame containing columns for `Total_CO2_Removed` and `Total_Lifecycle_Emissions`.}
}
\value{
A vector of NCRR values for each row in the dataset.
}
\description{
This function calculates the Net Carbon Removal Ratio (NCRR) for a given dataset.
}
\examples{
# Example usage:
calculate_ncrr(CDR_Lifecycle)
}
